refactor: isolate MCP dogfood jobs
